+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.11 (pre-release):
 * [Fix #248] Broken inventory for the new DLC suits.
+* [Fix #252] Allow setting type of the inventory via config.
 * [Fix #253] Kerbal preview is wrong for the expansion pack suites.
 
 # 1.10 (March 8th, 2018):